+++ src/JournalCli/Cmdlets/JournalCmdletBase.cs
@@ -1,6 +1,6 @@
 ﻿using System.Management.Automation;
 
-namespace JournalCli.Commands
+namespace JournalCli.Cmdlets
 {
     public abstract class JournalCmdletBase : CmdletBase
     {
