+++ src/JournalCli/Cmdlets/NewJournalEntryCmdlet.cs
@@ -4,7 +4,7 @@ using System.IO.Abstractions;
 using System.Management.Automation;
 using JetBrains.Annotations;
 
-namespace JournalCli.Commands
+namespace JournalCli.Cmdlets
 {
     [PublicAPI]
     [Cmdlet(VerbsCommon.New, "JournalEntry")]
