+++ src/JournalCli/Cmdlets/OpenBackupLocationCmdlet.cs
@@ -3,7 +3,7 @@ using System.IO.Abstractions;
 using System.Management.Automation;
 using JetBrains.Annotations;
 
-namespace JournalCli.Commands
+namespace JournalCli.Cmdlets
 {
     [PublicAPI]
     [Cmdlet(VerbsCommon.Open, "BackupLocation")]
