+++ src/JournalCli/Cmdlets/OpenRandomJournalEntryCmdlet.cs
@@ -2,7 +2,7 @@
 using System.Management.Automation;
 using JetBrains.Annotations;
 
-namespace JournalCli.Commands
+namespace JournalCli.Cmdlets
 {
     [PublicAPI]
     [Cmdlet(VerbsCommon.Open, "RandomJournalEntry")]
